Move `InMemoryBufferedSingleSignatureStore` to store module
To keep the `buffered_certifier` lean
