+++ mithril-aggregator/src/lib.rs
@@ -53,7 +53,8 @@ pub use snapshotter::{
     SnapshotterCompressionAlgorithm,
 };
 pub use store::{
-    CertificatePendingStore, ProtocolParametersStorer, VerificationKeyStore, VerificationKeyStorer,
+    CertificatePendingStore, InMemoryBufferedSingleSignatureStore, ProtocolParametersStorer,
+    VerificationKeyStore, VerificationKeyStorer,
 };
 pub use tools::{
     CExplorerSignerRetriever, SignersImporter, SignersImporterPersister, SignersImporterRetriever,
