+++ mithril-aggregator/src/store/mod.rs
@@ -1,7 +1,9 @@
+mod buffered_single_signatures_store;
 mod pending_certificate_store;
 mod protocol_parameters_store;
 mod verification_key_store;
 
+pub use buffered_single_signatures_store::InMemoryBufferedSingleSignatureStore;
 pub use pending_certificate_store::CertificatePendingStore;
 pub use protocol_parameters_store::ProtocolParametersStorer;
 pub use verification_key_store::{VerificationKeyStore, VerificationKeyStorer};
